feat: update package naming in benchmarks module, add module level IOUtils
