+++ benchmarks/src/jmh/kotlin/io/github/clasicrando/kdbc/benchmarks/postgresql/PgBenchmarkAsyncSingleJdbc.kt
@@ -1,4 +1,4 @@
-package com.github.kdbc.benchmarks.postgresql
+package io.github.clasicrando.kdbc.benchmarks.postgresql
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.runBlocking
