+++ benchmarks/src/jmh/kotlin/io/github/clasicrando/kdbc/benchmarks/postgresql/PgBenchmarkBlockingCopyKdbc.kt
@@ -1,9 +1,8 @@
-package com.github.kdbc.benchmarks.postgresql
+package io.github.clasicrando.kdbc.benchmarks.postgresql
 
-import io.github.clasicrando.kdbc.core.IOUtils
+import io.github.clasicrando.kdbc.benchmarks.IOUtils
 import io.github.clasicrando.kdbc.core.query.executeClosing
-import io.github.clasicrando.kdbc.postgresql.connection.PgAsyncConnection
-import kotlinx.coroutines.runBlocking
+import io.github.clasicrando.kdbc.postgresql.connection.PgBlockingConnection
 import kotlinx.io.buffered
 import kotlinx.io.files.Path
 import org.openjdk.jmh.annotations.Benchmark
@@ -27,10 +26,10 @@ import java.util.concurrent.TimeUnit
 @BenchmarkMode(Mode.AverageTime)
 @OutputTimeUnit(TimeUnit.MICROSECONDS)
 @State(Scope.Benchmark)
-open class PgBenchmarkAsyncCopyKdbc {
-    private val connection: PgAsyncConnection = runBlocking { getKdbcAsyncConnection() }
-    private val outputPath = Path(".", "temp", "kdbc-async-copy-out-benchmark.csv")
-    private val inputPath = Path(".", "temp", "kdbc-async-copy-in-benchmark.csv")
+open class PgBenchmarkBlockingCopyKdbc {
+    private val connection: PgBlockingConnection = getKdbcBlockingConnection()
+    private val outputPath = Path(".", "temp", "kdbc-blocking-copy-out-benchmark.csv")
+    private val inputPath = Path(".", "temp", "kdbc-blocking-copy-in-benchmark.csv")
     private val outputPathJava = java.nio.file.Path.of(outputPath.toString())
     private val inputPathJava = java.nio.file.Path.of(inputPath.toString())
 
@@ -38,16 +37,12 @@ open class PgBenchmarkAsyncCopyKdbc {
     open fun start() {
         IOUtils.createFileIfNotExists(outputPath)
         createBenchmarkCsv(inputPath)
-        runBlocking {
-            connection.createQuery(copyOutSetupQuery)
-                .executeClosing()
-            connection.createQuery(copyInSetupQuery)
-                .executeClosing()
-        }
+        connection.createQuery(copyOutSetupQuery).executeClosing()
+        connection.createQuery(copyInSetupQuery).executeClosing()
     }
 
     @Benchmark
-    open fun copyOutSink(): Unit = runBlocking {
+    open fun copyOutSink() {
         IOUtils.sink(outputPath).buffered().use { sink ->
             connection.copyOut(
                 copyOutStatement = kdbcCopyOut,
@@ -57,7 +52,7 @@ open class PgBenchmarkAsyncCopyKdbc {
     }
 
     @Benchmark
-    open fun copyOutStream(): Unit = runBlocking {
+    open fun copyOutStream() {
         Files.newOutputStream(outputPathJava).use { stream ->
             connection.copyOut(
                 copyOutStatement = kdbcCopyOut,
@@ -67,12 +62,12 @@ open class PgBenchmarkAsyncCopyKdbc {
     }
 
     @TearDown(Level.Invocation)
-    open fun cleanUp(): Unit = runBlocking {
+    open fun cleanUp() {
         connection.createQuery("TRUNCATE TABLE public.copy_in_posts;").executeClosing()
     }
 
     @Benchmark
-    open fun copyInSource(): Unit = runBlocking {
+    open fun copyInSource() {
         IOUtils.source(inputPath).buffered().use { source ->
             connection.copyIn(
                 copyInStatement = kdbcCopyIn,
@@ -82,7 +77,7 @@ open class PgBenchmarkAsyncCopyKdbc {
     }
 
     @Benchmark
-    open fun copyInStream(): Unit = runBlocking {
+    open fun copyInStream() {
         Files.newInputStream(inputPathJava).use { stream ->
             connection.copyIn(
                 copyInStatement = kdbcCopyIn,
@@ -95,13 +90,11 @@ open class PgBenchmarkAsyncCopyKdbc {
     fun destroy() {
         IOUtils.deleteCatching(path = outputPath, mustExist = false)
         IOUtils.deleteCatching(path = inputPath, mustExist = false)
-        runBlocking {
-            if (connection.isConnected) {
-                try {
-                    connection.close()
-                } catch (ex: Throwable) {
-                    ex.printStackTrace()
-                }
+        if (connection.isConnected) {
+            try {
+                connection.close()
+            } catch (ex: Throwable) {
+                ex.printStackTrace()
             }
         }
     }
