+++ benchmarks/src/jmh/kotlin/io/github/clasicrando/kdbc/benchmarks/postgresql/PgBenchmarkBlockingMultiKdbc.kt
@@ -1,5 +1,9 @@
-package com.github.kdbc.benchmarks.postgresql
+package io.github.clasicrando.kdbc.benchmarks.postgresql
 
+import io.github.clasicrando.kdbc.core.query.bind
+import io.github.clasicrando.kdbc.core.query.executeClosing
+import io.github.clasicrando.kdbc.core.query.fetchAll
+import io.github.clasicrando.kdbc.postgresql.pool.PgBlockingConnectionPool
 import org.openjdk.jmh.annotations.Benchmark
 import org.openjdk.jmh.annotations.BenchmarkMode
 import org.openjdk.jmh.annotations.Fork
@@ -14,7 +18,6 @@ import java.util.concurrent.Executor
 import java.util.concurrent.ExecutorCompletionService
 import java.util.concurrent.Executors
 import java.util.concurrent.TimeUnit
-import javax.sql.DataSource
 
 @Warmup(iterations = 4, time = 10, timeUnit = TimeUnit.SECONDS)
 @Measurement(iterations = 20, time = 10, timeUnit = TimeUnit.SECONDS)
@@ -22,35 +25,34 @@ import javax.sql.DataSource
 @BenchmarkMode(Mode.AverageTime)
 @OutputTimeUnit(TimeUnit.MICROSECONDS)
 @State(Scope.Benchmark)
-open class PgBenchmarkBlockingMultiJdbc {
+open class PgBenchmarkBlockingMultiKdbc {
     private var id = 0
-    private val dataSource: DataSource = getJdbcDataSource()
+    private val pool = PgBlockingConnectionPool(
+        connectOptions = kdbcConnectOptions,
+        poolOptions = poolOptions,
+    )
     private val executor: Executor = Executors.newWorkStealingPool()
     private val completionService = ExecutorCompletionService<Unit>(executor)
 
     @Setup
     open fun start() {
-        getJdbcConnection().use { connection ->
-            connection.createStatement().use { statement ->
-                statement.execute(setupQuery)
-            }
+        pool.acquire().use {
+            it.createQuery(setupQuery)
+                .executeClosing()
         }
     }
 
-    private fun singleStep(): Int {
+    private fun step(): Int {
         id++
         if (id > 5000) id = 1
         return id
     }
 
-    private fun executeQuery(stepId: Int) {
-        dataSource.connection.use { conn ->
-            conn.prepareStatement(jdbcQuerySingle).use { preparedStatement ->
-                preparedStatement.setInt(1, stepId)
-                preparedStatement.executeQuery().use { resultSet ->
-                    extractPostDataClassListFromResultSet(resultSet)
-                }
-            }
+    private fun executeQuery(stepId: Int): List<PostDataClass> {
+        return pool.acquire().use { conn ->
+            conn.createPreparedQuery(kdbcQuerySingle)
+                .bind(stepId)
+                .fetchAll(PostDataClassRowParser)
         }
     }
 
@@ -58,7 +60,7 @@ open class PgBenchmarkBlockingMultiJdbc {
     open fun querySingleRow() {
         val taskCount = concurrencyLimit
         repeat(taskCount) {
-            val stepId = singleStep()
+            val stepId = step()
             completionService.submit { executeQuery(stepId) }
         }
         var received = 0
