+++ benchmarks/src/jmh/kotlin/io/github/clasicrando/kdbc/benchmarks/postgresql/PgBenchmarkBlockingSingleKdbc.kt
@@ -1,10 +1,9 @@
-package com.github.kdbc.benchmarks.postgresql
+package io.github.clasicrando.kdbc.benchmarks.postgresql
 
-import io.github.clasicrando.kdbc.core.connection.AsyncConnection
+import io.github.clasicrando.kdbc.core.connection.BlockingConnection
 import io.github.clasicrando.kdbc.core.query.bind
 import io.github.clasicrando.kdbc.core.query.executeClosing
 import io.github.clasicrando.kdbc.core.query.fetchAll
-import kotlinx.coroutines.runBlocking
 import org.openjdk.jmh.annotations.Benchmark
 import org.openjdk.jmh.annotations.BenchmarkMode
 import org.openjdk.jmh.annotations.Fork
@@ -24,13 +23,14 @@ import java.util.concurrent.TimeUnit
 @BenchmarkMode(Mode.AverageTime)
 @OutputTimeUnit(TimeUnit.MICROSECONDS)
 @State(Scope.Benchmark)
-open class PgBenchmarkAsyncSingleKdbc {
+open class PgBenchmarkBlockingSingleKdbc {
     private var id = 0
-    private val connection: AsyncConnection = runBlocking { getKdbcAsyncConnection() }
+    private val connection: BlockingConnection = getKdbcBlockingConnection()
 
     @Setup
-    open fun start(): Unit = runBlocking {
-        connection.createQuery(setupQuery).executeClosing()
+    open fun start() {
+        connection.createQuery(setupQuery)
+            .executeClosing()
     }
 
     private fun singleStep(): Int {
@@ -45,7 +45,7 @@ open class PgBenchmarkAsyncSingleKdbc {
     }
 
     @Benchmark
-    open fun querySingleRow(): Unit = runBlocking {
+    open fun querySingleRow() {
         singleStep()
         connection.createPreparedQuery(kdbcQuerySingle)
             .bind(id)
@@ -53,7 +53,7 @@ open class PgBenchmarkAsyncSingleKdbc {
     }
 
     @Benchmark
-    open fun queryMultipleRows(): Unit = runBlocking {
+    open fun queryMultipleRows() {
         multiStep()
         connection.createPreparedQuery(kdbcQuery)
             .bind(id)
@@ -63,13 +63,11 @@ open class PgBenchmarkAsyncSingleKdbc {
 
     @TearDown
     fun destroy() {
-        runBlocking {
-            if (connection.isConnected) {
-                try {
-                    connection.close()
-                } catch (ex: Throwable) {
-                    ex.printStackTrace()
-                }
+        if (connection.isConnected) {
+            try {
+                connection.close()
+            } catch (ex: Throwable) {
+                ex.printStackTrace()
             }
         }
     }
